Initl change of placing all files in one big index.js file
