+++ index.html
@@ -19,6 +19,6 @@
         <img id="imageBrick" src="assets/images/brick.jpg">
 
         <canvas id="gameScreen" width="800" height="600"></canvas>
-        <script type="module" src="src/index.js"></script>
+        <script type="module" src="index.js"></script>
     </body>
 </html>
