--- src/ball.js
@@ -1,50 +1,0 @@
-import { detectCollision } from '/src/collisionDetection.js';
-
-export default class Ball {
-    constructor(game) {
-        this.image = document.getElementById("imageBall");
-
-        this.gameWidth = game.gameWidth;
-        this.gameHeight = game.gameHeight;
-
-        this.game = game;
-        this.reset();
-
-        this.size = 16;
-    }
-
-    reset() {
-        this.position = {x: 10, y: 400}
-        this.speed = {x: 4, y: -2}
-    }
-
-    draw(ctx) {
-        ctx.drawImage(this.image, this.position.x, this.position.y, this.size, this.size);
-    }
-
-    update(deltatime) {
-        this.position.x += this.speed.x;
-        this.position.y += this.speed.y;
-
-        // Wall on left or right
-        if (this.position.x + this.size > this.gameWidth || this.position.x < 0) {
-            this.speed.x = -this.speed.x;
-        }
-
-        // wall on top
-        if (this.position.y < 0) {
-            this.speed.y = -this.speed.y;
-        }
-
-        // Bottom of game
-        if (this.position.y + this.size > this.gameHeight) {
-            this.game.lives--;
-            this.reset();
-        }
-
-        if (detectCollision(this, this.game.paddle)) {
-            this.speed.y = -this.speed.y;
-            this.position.y = this.game.paddle.position.y - this.size;
-        }
-    }
-}
