--- src/game.js
@@ -1,131 +1,0 @@
-import Paddle from '/src/paddle.js';
-import InputHandler from '/src/input.js';
-import Ball from '/src/ball.js';
-import Brick from '/src/brick.js';
-
-import { buildLevel, level1, level2 } from '/src/levels.js';
-
-const GAMESTATE = {
-    PAUSED: 0,
-    RUNNING: 1,
-    MENU: 2,
-    GAMEOVER: 3,
-    NEWLEVEL: 4,
-    WIN: 5
-};
-
-export default class Game {
-    constructor(gameWidth, gameHeight) {
-
-        this.gameWidth = gameWidth;
-        this.gameHeight = gameHeight;
-        this.gamestate = GAMESTATE.MENU;
-        this.paddle = new Paddle(this);
-        this.ball = new Ball(this);
-
-        this.gameObjects = [];
-        this.bricks = [];
-        this.lives = 3;
-
-        this.levels = [level1, level2];
-        this.currentLevel = 0;
-
-        new InputHandler(this.paddle, this);
-    }
-
-    start() {
-
-        if (this.gamestate !== GAMESTATE.MENU && this.gamestate !== GAMESTATE.NEWLEVEL) return;
-
-        this.bricks = buildLevel(this, this.levels[this.currentLevel]);
-        this.ball.reset();
-
-        this.gameObjects = [
-            this.ball,
-            this.paddle
-        ];
-
-        this.gamestate = GAMESTATE.RUNNING;
-    }
-
-    update(deltaTime) {
-
-        if (this.lives == 0) this.gamestate = GAMESTATE.GAMEOVER;
-
-        if (this.gamestate === GAMESTATE.PAUSED || this.gamestate === GAMESTATE.MENU || this.gamestate === GAMESTATE.GAMEOVER) return;
-
-        if (this.bricks.length === 0) {
-            this.currentLevel++;
-            if (this.currentLevel < this.levels.length) {
-                this.gamestate = GAMESTATE.NEWLEVEL;
-                this.start();
-            }
-            else {
-                this.gamestate = GAMESTATE.WIN;
-            }
-        }
-
-        [...this.gameObjects, ...this.bricks].forEach((object) => object.update(deltaTime));
-
-        this.bricks = this.bricks.filter(brick => !brick.toBeDestroyed);
-    }
-
-    draw(ctx) {
-        [...this.gameObjects, ...this.bricks].forEach((object) => object.draw(ctx));
-
-        if (this.gamestate === GAMESTATE.PAUSED) {
-            ctx.rect(0, 0, this.gameWidth, this.gameHeight);
-            ctx.fillStyle = "rgba(0, 0, 0, 0.5)";
-            ctx.fill();
-
-            ctx.font = "30px Arial";
-            ctx.fillStyle = "white";
-            ctx.textAlign = "center";
-            ctx.fillText("Paused", this.gameWidth / 2, this.gameHeight / 2);
-        }
-
-        if (this.gamestate === GAMESTATE.MENU) {
-            ctx.rect(0, 0, this.gameWidth, this.gameHeight);
-            ctx.fillStyle = "black";
-            ctx.fill();
-
-            ctx.font = "30px Arial";
-            ctx.fillStyle = "white";
-            ctx.textAlign = "center";
-            ctx.fillText("Press Spacebar To Begin", this.gameWidth / 2, this.gameHeight / 2);
-        }
-
-        if (this.gamestate === GAMESTATE.GAMEOVER) {
-            ctx.rect(0, 0, this.gameWidth, this.gameHeight);
-            ctx.fillStyle = "black";
-            ctx.fill();
-
-            ctx.font = "30px Arial";
-            ctx.fillStyle = "white";
-            ctx.textAlign = "center";
-            ctx.fillText("You've met with a terrible fate, haven't you?", this.gameWidth / 2, this.gameHeight / 2);
-        }
-
-        if (this.gamestate === GAMESTATE.WIN) {
-            ctx.rect(0, 0, this.gameWidth, this.gameHeight);
-            ctx.fillStyle = "red";
-            ctx.fill();
-
-            ctx.font = "30px Arial";
-            ctx.fillStyle = "green";
-            ctx.textAlign = "center";
-            ctx.fillText("You Win!", this.gameWidth / 2, this.gameHeight / 2);
-        }
-    }
-
-    togglePause() {
-        if (this.gamestate !== GAMESTATE.MENU) {
-            if (this.gamestate == GAMESTATE.PAUSED) {
-                this.gamestate = GAMESTATE.RUNNING;
-            }
-            else {
-                this.gamestate = GAMESTATE.PAUSED;
-            }
-        }
-    }
-}
